requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-decouple==3.8

# Database
motor==3.3.2
beanie==1.24.0
pymongo==4.6.1

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Plotting
plotly==5.18.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Optional: Production server
# gunicorn==21.2.0

